requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
redis==5.2.1
celery==5.5.0
requests==2.32.3
mongoengine==0.29.1
pymongo==4.11.3
python-dotenv==1.0.0
pydantic==2.3.0
parsel